scripts/requirements.txt
alabaster==0.7.12
aniso8601==9.0.1
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==0.14.7
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.1.0
bleach==4.1.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.1
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.1.0
docutils==0.17.1
elasticsearch==7.17.0
elasticsearch-dsl==7.4.0
email-validator==1.1.3
entrypoints==0.3
executing==0.8.2
Flask==2.0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
ghp-import==2.0.1
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
greenlet==1.1.1
h11==0.13.0
httplib2==0.19.1
icalendar==4.0.7
ics==0.7
idna==3.2
imagesize==1.2.0
importlib-metadata==4.6.1
iniconfig==1.1.1
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
jwt==1.3.1
livereload==2.6.3
Markdown==3.3.7
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
matplotlib-inline==0.1.3
mergedeep==1.3.4
mistune==0.8.4
mkdocs==1.3.0
mkdocs-material==2.9.4
mkdocs-material-extensions==1.0.1
mkdocs-windmill==0.1.6
mypy-extensions==0.4.3
mysql-connector-python==8.0.26
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.20.3
oauthlib==3.1.1
outcome==1.1.0
packaging==21.0
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
pluggy==1.0.0
postmarker==0.18.2
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==3.17.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.9.0
pymdown-extensions==5.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-frontmatter==1.0.0
pytz==2021.1
PyYAML==5.4.1
pyyaml-env-tag==0.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.9.2
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
selenium==4.1.5
Send2Trash==1.8.0
shyaml==0.6.2
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==4.2.0
sphinx-autodoc-typehints==1.12.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.23
stack-data==0.1.4
TatSu==4.4.0
terminado==0.13.1
testpath==0.5.0
tomli==2.0.0
tornado==4.5.3
tqdm==4.64.0
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.6
uuid==1.30
watchdog==2.1.3
wcwidth==0.2.5
webdriverdownloader==1.1.0.3
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.2
wsproto==1.1.0
WTForms==2.3.3
zipp==3.5.0
